refactor: separate flat config
