+++ src/index.ts
@@ -1,66 +1,10 @@
-import * as parserJsonc from "jsonc-eslint-parser";
-import { createRequire } from "node:module";
+import { configs as recommendedConfigs } from "./configs/recommended.js";
+import { plugin } from "./plugin.js";
 
-import { rule as orderProperties } from "./rules/order-properties.js";
-import { rule as preferRepositoryShorthand } from "./rules/prefer-repository-shorthand.js";
-import { rule as sortCollections } from "./rules/sort-collections.js";
-import { rule as uniqueDependencies } from "./rules/unique-dependencies.js";
-import { rule as validLocalDependency } from "./rules/valid-local-dependency.js";
-import { rule as validName } from "./rules/valid-name.js";
-import { rule as validPackageDef } from "./rules/valid-package-def.js";
-import { rule as validRepositoryDirectory } from "./rules/valid-repository-directory.js";
-import { rule as validVersion } from "./rules/valid-version.js";
-
-const require = createRequire(import.meta.url || __filename);
-
-const { name, version } = require("../package.json") as {
-	name: string;
-	version: string;
-};
-
-export const rules = {
-	"order-properties": orderProperties,
-	"prefer-repository-shorthand": preferRepositoryShorthand,
-	"sort-collections": sortCollections,
-	"unique-dependencies": uniqueDependencies,
-	"valid-local-dependency": validLocalDependency,
-	"valid-name": validName,
-	"valid-package-def": validPackageDef,
-	"valid-repository-directory": validRepositoryDirectory,
-	"valid-version": validVersion,
-};
-
-const recommendedRuleSettings = Object.fromEntries(
-	Object.entries(rules)
-		.filter(([, rule]) => rule.meta.docs?.recommended)
-		.map(([name]) => ["package-json/" + name, "error" as const]),
-);
-
-const plugin = {
-	configs: {
-		recommended: {
-			files: ["**/package.json"],
-			languageOptions: {
-				parser: parserJsonc,
-			},
-			plugins: {
-				get "package-json"() {
-					return plugin;
-				},
-			},
-			rules: recommendedRuleSettings,
-		},
-		"recommended-legacy": {
-			rules: recommendedRuleSettings,
-		},
+export const rules = plugin.rules;
+export const configs = {
+	recommended: {
+		plugins: ["package-json"],
+		rules: recommendedConfigs.recommended.rules,
 	},
-	meta: {
-		name,
-		version,
-	},
-	rules,
 };
-
-export const configs = plugin.configs;
-
-export default plugin;
